+++ crates/frontend/src/components/dimension_form/types.rs
@@ -4,7 +4,7 @@ use serde_json::Value;
 #[derive(Serialize, Clone)]
 pub struct DimensionCreateReq {
     pub dimension: String,
-    pub priority: u16,
+    pub priority: u32,
     pub schema: Value,
     pub function_name: Option<Value>,
 }
